+++ db/migrate/20190403204257_remove_phone_from_attendances.rb
@@ -1,0 +1,5 @@
+class RemovePhoneFromAttendances < ActiveRecord::Migration[5.2]
+  def change
+    remove_column :attendances, :phone, :string
+  end
+end