+++ spec/factories/attendance.rb
@@ -7,7 +7,6 @@ FactoryBot.define do
     sequence(:email) { |n| "email#{n}@email.com" }
     net_id { 'abc123' }
     upi { '01234567' }
-    phone { '555-555-5555' }
     check_in { Time.zone.now }
     event
   end