+++ spec/rails_helper.rb
@@ -64,7 +64,7 @@ RSpec.configure do |config|
   config.filter_rails_from_backtrace!
   # arbitrary gems may also be filtered via:
   # config.filter_gems_from_backtrace("gem name")
-
+  
   config.include Devise::Test::IntegrationHelpers, type: :system
 
   config.before(:each, type: :system) do